requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
httpx==0.24.0
openai==0.14.0
PyPDF2==3.0.0
python-dotenv==0.21.0